refactor: move some update logic to belts.rs and pipes.rs
